+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [2.1.0] - 2020-10-10
+### Added
+* Additional error handling
+* New tests
+
+### Changes
+* New max dimensions, to stop heap out of memory errors
+* Minor documentation fixes
 
 ## [2.0.0] - 2020-04-21
 ### Added